🚚 Move the WindowSize to common module
Once this new class will be used in more than one Compose-module, the
code was moved to Design System by now. In the future, this class will
be available in the WindowManager library and it could be removed.
