+++ app/src/androidTest/java/com/escodro/alkaa/HomeScreenTest.kt
@@ -8,6 +8,7 @@ import androidx.test.platform.app.InstrumentationRegistry
 import com.escodro.alkaa.model.HomeSection
 import com.escodro.alkaa.navigation.NavGraph
 import com.escodro.designsystem.AlkaaTheme
+import com.escodro.designsystem.WindowSize
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
